+++ snippets/footer.php
@@ -3,5 +3,13 @@
     Entypo pictograms by Daniel Bruce - <a href="http://www.entypo.com">www.entypo.com</a>
 </footer>
 
+<?= js('/assets/plugins/kirby-serp/js/dist/script.min.js'); ?>
+
+<script>
+document.addEventListener('DOMContentLoaded', function() {
+	affiliateBeautifier.init();
+});
+</script>
+
 </body>
 </html>