+++ snippets/meta.php
@@ -1,0 +1,4 @@
+<div class="meta">
+    <div class="edit"><a href=""></a></div>
+    <div class="panel"><a href=""></a></div>
+</div>