+++ docs/changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+**0.2**
+
+- Click to open navigation for each item instead of hover.
+- Config option for title and description fallbacks.
+- Navigation tabs is added and includes home and warnings so far.
+- Removed `plugin.serp.collection` and replaced it with `plugin.serp.filter.query` and it now allows filter logic.
+- Changed flag names from `warning` and `danger` to `major` and `minor`.
+
 **0.1**
 
 - Initial release